+++ logicmonitor/resource_device_group_test.go
@@ -13,51 +13,23 @@ import (
 
 func TestAccLogicMonitorDeviceGroup(t *testing.T) {
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testDeviceGroupDestroy,
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCheckLogicMonitorDeviceGroup,
 				Check: resource.ComposeTestCheckFunc(
 					testDeviceGroupExists("logicmonitor_device_group.group1"),
 					resource.TestCheckResourceAttr(
-						"logicmonitor_device_group.group1", "name", "NewGroup"),
+						"logicmonitor_device_group.group1", "name", "BlackMamba"),
 					resource.TestCheckResourceAttr(
 						"logicmonitor_device_group.group1", "disable_alerting", "true"),
 					resource.TestCheckResourceAttr(
-						"logicmonitor_device_group.group1", "properties.group", "test"),
-					resource.TestCheckResourceAttr(
-						"logicmonitor_device_group.group1", "properties.system.categories", "a,b,c,d"),
+						"logicmonitor_device_group.group1", "properties.system.categories", "GOAT,MVP"),
 				),
 			},
 		},
 	})
-}
-
-func testDeviceGroupDestroy(s *terraform.State) error {
-	client := testAccProvider.Meta().(*lmclient.LMSdkGo)
-	if err := testCollectorGroupDestroyHelper(s, client); err != nil {
-		return err
-	}
-	return nil
-}
-
-func testDeviceGroupDestroyHelper(s *terraform.State, client *lmclient.LMSdkGo) error {
-	for _, r := range s.RootModule().Resources {
-		id, e := strconv.Atoi(r.Primary.ID)
-		if e != nil {
-			return e
-		}
-		params := lm.NewDeleteDeviceGroupByIDParams()
-		params.SetID(int32(id))
-
-		nil, err := client.LM.DeleteDeviceGroupByID(params)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 func testDeviceGroupExists(n string) resource.TestCheckFunc {
@@ -90,13 +62,12 @@ func testDeviceGroupExistsHelper(s *terraform.State, client *lmclient.LMSdkGo) e
 
 const testAccCheckLogicMonitorDeviceGroup = `
 resource "logicmonitor_device_group" "group1" {
-    name = "NewGroup"
+    name = "BlackMamba"
     disable_alerting = true
     description = "testing group"
     applies_to = "system.displayname =~ \"test\""
     properties = {
-     "group" = "test"
-     "system.categories" = "a,b,c,d"
+     "system.categories" = "GOAT,MVP"
     }
 }
 `